scaffold main CLI options
