+++ ol/genesis-tools/src/live_fork.rs
@@ -4,29 +4,31 @@ use crate::fetch_archive::wait_for_archive;
 use anyhow::Result;
 
 /// starts the daemon for the live fork.
-pub fn fork_daemon() {
-  match wait_for_archive() {
-    Ok(_) => {
-      // generate_genesis_from_snapshot
-      restart()
-    },
-    Err(_) => todo!(),
+pub fn fork_daemon() -> Result<()> {
+    match wait_for_archive() {
+        Ok(_) => {
+            // generate_genesis_from_snapshot
+            match restart() {
+                Ok(_) => Ok(()),
+                Err(e) => Err(e),
+            }
+        }
+        Err(e) => Err(e),
+    }
 }
-}
-
-
 
 /// restart node from new genesis.blob from archive
-pub fn restart() -> Result<()>{
-  kill_all_services();
+pub fn restart() -> Result<()> {
+    kill_all_services();
 
-  ol_start();
+    ol_start();
+    Ok(())
 }
 
 fn kill_all_services() {
-  todo!()
+    todo!()
 }
 
 fn ol_start() {
-  todo!()
+    todo!()
 }
